fix: Windows build warnings - format specifiers and pragmas (v0.3.54) [BUILD]
Format specifier fixes for uint64_t/time_t on Windows:
- dht_contact_request.c, dna_feed_posts.c, dna_feed_comments.c
- transport_helpers.c (time_t)

Pragma comment fixes (MSVC-only):
- dht_contactlist.c, dht_message_backup.c, transport_core.h

Other fixes:
- keygen.c: Include winsock2.h before windows.h
- transport_discovery.c: Fix unused variable on Windows
- win32/getopt.c: Remove unused nextchar variable

+++ p2p/transport/transport_helpers.c
@@ -267,8 +267,8 @@ int stun_get_public_ip(char *ip_out, size_t len) {
  */
 int create_presence_json(const char *ips, uint16_t port, char *json_out, size_t len) {
     int written = snprintf(json_out, len,
-                          "{\"ips\":\"%s\",\"port\":%d,\"timestamp\":%ld}",
-                          ips, port, time(NULL));
+                          "{\"ips\":\"%s\",\"port\":%d,\"timestamp\":%lld}",
+                          ips, port, (long long)time(NULL));
     return (written >= 0 && (size_t)written < len) ? 0 : -1;
 }
